+++ src/JohnStevenson/JsonWorks/Schema/ValidationException.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace JohnStevenson\JsonWorks\Schema;
+
+class ValidationException extends \Exception {}