+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Recent changes not in tagged release
 * Fixed newline bug in Utils:dataToJson
+* Added JohnStevenson\JsonWorks\Schema\ValidationException
 
 ## v1.0.0
 [2013-04-22](https://github.com/johnstevenson/json-works/tree/v1.0.0)